Restore StructTests
