+++ tests/IceRpc.Slice.Tests/Slice1NullableTests.cs
@@ -1,6 +1,5 @@
 // Copyright (c) ZeroC, Inc.
 
-using IceRpc.Tests.Common;
 using NUnit.Framework;
 using ZeroC.Slice.Codec;
 using ZeroC.Tests.Common;
